+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 
+* Fix `Images::prune` return type.
+
 # 0.2.2
 * Fix `port_bindings` field of `InspectContainerHostConfig` and `InspectPodInfraConfig`. The value in the hashmap can be null.
 * Fix `PortMap` type signature.