+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#
+* Add `portmappings` to `ContainerCreateOptsBuilder` [#117](https://github.com/vv9k/podman-api-rs/pull/117)
+
 # 0.2.3 
 * Fix `Images::prune` return type.