+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #
 * Add `portmappings` to `ContainerCreateOptsBuilder` [#117](https://github.com/vv9k/podman-api-rs/pull/117)
+* Fix typo `referene` in `PullOptsBuilder` [#119](https://github.com/vv9k/podman-api-rs/pull/119)
+
 
 # 0.2.3 
 * Fix `Images::prune` return type.