+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Add `portmappings` to `ContainerCreateOptsBuilder` [#117](https://github.com/vv9k/podman-api-rs/pull/117)
 * Fix typo `referene` in `PullOptsBuilder` [#119](https://github.com/vv9k/podman-api-rs/pull/119)
 * Make Images::pull return a stream [#121](https://github.com/vv9k/podman-api-rs/pull/121)
+* *BREAKING* All API structs no longer have a `'podman` lifetime. This change makes it easier to create self working objects without
+  the lifetime hell and according to hyper client documentation it is cheap to clone and cloning is the recommended way to share a client.
 
 
 # 0.2.3